fix(tests): 2 more red-main fixes surfaced by the same CI run
1. MarathonEnrollmentTest::test_content_falls_back_to_shared_default_
   when_cohort_has_no_override used `day1_message` as its no-override
   example -- true when H547 wrote it, but H445 Phase 3 (PR #446, same
   run) gave `deva` a real day1_message override, breaking the
   assumption. Switched to `day3_message_paid`, which is
   design-permanently shared across cohorts (host/schedule never vary
   by cohort per config/marathon.php's own Phase 1 comment).

2. ClassAttendanceServiceTest expected an int rate (assertSame(100, ...))
   but ClassAttendanceService::$withRate computed round(...) without a
   cast, which PHP always returns as float -- unrelated to marathon
   (H553's attendance dashboard), but same CI run, same fix-forward
   principle as the day5 off-by-one (H437). Cast to int at the
   computation site: a whole-number attendance percentage is
   semantically an int, not a float artifact of round().

+++ app/Services/ClassAttendanceService.php
@@ -200,7 +200,7 @@ class ClassAttendanceService
         }
 
         $withRate = fn (array $row): array => $row + [
-            'rate' => $row['expected'] > 0 ? round($row['attended'] / $row['expected'] * 100) : 0,
+            'rate' => $row['expected'] > 0 ? (int) round($row['attended'] / $row['expected'] * 100) : 0,
         ];
 
         // Хронические неявки: последние $chronicThreshold занятий подряд (history
